Shuffleboard Manual Pattern Input
The changes outside of 71-107 of Drive Command were reformatting things, just making spacing and visuals consistent, I accidentally clicked something that reformats everything but it looks a lot better now and it worked when testing with simulation. Json stuff was because of shuffleboard, I didn't touch anything personally but it's there.  Added a solid black constant, should be used for when errors or issues arise with the lights. fixed an inconsistency in blinkin subsystem.

Changes From 71-107 in Drive Command (The Actual Important Changes):

Solved issue with code not recognizing input from shuffleboard, it was searching for a double where there was a string so returned only default. Now gets string and turns to double for pattern code.

Commented Line making stationary white, that should still be used when in normal operations. When robot is stationary, takes a raw pattern code that can be inputted from shuffleboard, various conditions and adjustments get made to make sure it will work as an actual code or it wont break everything if it isnt.

Adjustments change any inputted decimal (or 0) of any length into a 2 digit decimal that is one of the codes, rounds down if in between.

If something other than a decimal is inputted (ie. letter, whole number, nothing), catches exception and sets leds to solid black, which will be our designated light code for errors, can be used in other things.

Everything's been tested and worked fine in simulator, only potential issue is  when using simulator + shuffleboard, reactivating the simulator will cause you to need to manually update code in simulator networktables shuffleboard section. I don't know how this would work on the actual robot.

Program does need robot testing